// src/api/fiberlaser/models/ListPlatesDeliveryResponseDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { IdentifierPlateDto } from "./IdentifierPlateDto";
import type { PartNameDto } from "./PartNameDto";

export enum ListPlatesDeliveryResponseDtoDeliveryStatusEnum {
    delivered = "delivered",
    pending = "pending"
}

export type ListPlatesDeliveryResponseDto = {
    /**
     * @type number
    */
    ProductionID: number;
    /**
     * @type string
    */
    productName: string;
    /**
     * @type number
    */
    priority: number;
    /**
     * @type string
    */
    OrderNum: string;
    /**
     * @type string
    */
    PartCode: string;
    /**
     * @type object
    */
    PartName: PartNameDto;
    /**
     * @type number
    */
    PlanQty: number;
    /**
     * @type string
    */
    PlannedEndTimestamp: string;
    /**
     * @type array
    */
    Identifiersplates: IdentifierPlateDto[];
    /**
     * @type string
    */
    deliveryStatus: ListPlatesDeliveryResponseDtoDeliveryStatusEnum;
    /**
     * @type number
    */
    PlatesDeliveryID?: number | null;
    /**
     * @type string
    */
    Separator?: string | null;
    /**
     * @type string
    */
    deliveredData?: string | null;
};